Changed most of the parameters to psdsim() into doubles. Changed some betad[i][j] to betd (since i hasn't been initialized at that point. Continued porting main Fortran a little past the first call to psdsim()

+++ ctest.cpp
@@ -83,7 +83,7 @@ int main()
 	// DUMMY COMMENT
 
   // Now test psdsim
-  float spsd[BLZSIM_DIM16384]; // where to put the output light curve
+  double spsd[BLZSIM_DIM16384]; // where to put the output light curve
   float psdslp = 1.7; // psd slope
   float tinc= 1.93893981; 
   BlzSim* pSim = new BlzSim();
